Adjust link for DoP button
The Deploy on Platform.sh link was still pointing to the template maintained by Platform.sh, not this repo.  Assuming you want it to be this repo, here's a PR to do so.

Though I'm curious what you're changing in this repo that makes it better to use than the template?  Perhaps we can just update the template?